refactor: simplify clinic country enforcement
